Updated git script

+++ repo.sh
@@ -2,47 +2,63 @@
 
 echo "-----------Publish Commits------------"
 
-# Initialize the repository if needed
-if ! [ -d .git ]; then
-    git init
-    echo "The repo is now initialized"
+while true; do
 
-fi
+    # Initialize the repository if needed
+    read -p "Publish repo? [y/n] " option
 
-#Initialize gitignore if needed
-if ! [ -e .gitignore ]; then
-    touch .gitignore
-fi
+    if [[ "$option" == "y" ]]; then
 
-if [ ! -s .gitignore ]; then
-    echo -e "# General\n.DS_Store\n.AppleDouble" > .gitignore
-    echo "gitignore successfully generated"
+        if ! [ -d .git ]; then
+            git init
+            echo "The repo is now initialized"
 
-fi
+        fi
 
-#Create remote if needed
-read -p "Does remote exist? [n to enter address] " option
+        #Initialize gitignore if needed
+        if ! [ -e .gitignore ]; then
+            touch .gitignore
+        fi
 
-if [ "$option" == "n" ]; then
-    read -p "Enter an address for the repo: " repo_address
+        if [ ! -s .gitignore ]; then
+            echo -e "# General\n.DS_Store\n.AppleDouble" > .gitignore
+            echo "gitignore successfully generated"
 
-    git remote add origin "$repo_address"
+        fi
 
-fi
+        #Create remote if needed
+        read -p "Does remote exist? [n to enter address] " option
+
+        if [ "$option" == "n" ]; then
+            read -p "Enter an address for the repo: " repo_address
+
+            git remote add origin "$repo_address"
+
+        fi
+
+        #Stage changes and publish repo
+        git add .
+        echo "Changes staged."
 
 
-#Stage changes and publish repo
-git add .
-echo "Changes staged."
+        read -p "Enter commit message: " commit_message
+
+        git commit -m "$commit_message"
+
+        read -p "Enter branch name: " branch_name
+
+        git push origin "$branch_name"
+        echo "Commit published successfully"
+
+    #Cancels creating any git files or publishing the repo
+    elif [[ "$option" == "n" ]]; then
+        echo "Publishing repo canceled."
+        exit 0
+
+    else
+        echo "Please enter [y/n]"
+    fi
+done
 
 
-read -p "Enter commit message: " commit_message
-
-git commit -m "$commit_message"
-
-read -p "Enter branch name: " branch_name
-
-git push origin "$branch_name"
-
-echo "Commit published successfully"
